Change `MSImage` to `NSImage` (#63)
I'm not sure if there was ever an `MSImage` to use, but there are no docs for this class, and in my plugin, I used an instance of:

- NSImage
- NSBitmapImageRep

and both worked as far as I can tell. I'm assuming this was supposed to be `NSImage` here ? or some generic thing?